+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+## 1.0.3
+* Release date: 30/07/2024
+* Update display photo with custom photo container
+
+## 1.0.2
+* Release date: 25/07/2024
+* Update README with sample images
+
 ## 1.0.1
-* Fix issue with type grid5 (25/07/2024)
+* Release date: 25/07/2024
+* Fix issue with type grid5
 
 ## 1.0.0
-* First release (25/07/2024)
+* Release date: 25/07/2024
+* First release